refactor: Update utilities import in multiple scripts
