+++ scripts/postprocessing.py
@@ -14,7 +14,7 @@ sys.path.append(str(dockm8_path))
 
 from scripts.consensus_methods import CONSENSUS_METHODS
 from scripts.rescoring_functions import RESCORING_FUNCTIONS
-from scripts.utilities import printlog
+from scripts.utilities.utilities import printlog
 
 warnings.filterwarnings("ignore", category=UserWarning)
 warnings.filterwarnings("ignore", category=DeprecationWarning)
